[PL] extracted Python BC cmake target
