+++ ProcessLib/BoundaryCondition/Python/PythonBoundaryConditionPythonSideInterface.h
@@ -9,15 +9,8 @@
 
 #pragma once
 
-#include <pybind11/pybind11.h>
-
 namespace ProcessLib
 {
-//! Can be thrown to indicate that a member function is not overridden in a
-//! derived class (in particular, if a Python class inherits from a C++ class).
-struct MethodNotOverriddenInDerivedClassException
-{
-};
 //! Base class for boundary conditions.
 //! This class will get Python bindings and is intended to be to be derived in
 //! Python.
@@ -78,34 +71,5 @@ private:
     mutable bool _overridden_essential = true;
     //! Tells if getFlux() has been overridden in the derived class in Python.
     mutable bool _overridden_natural = true;
-};
-
-//! Trampoline class allowing the the abstract base class to be overridden.
-//! Cf. https://pybind11.readthedocs.io/en/stable/advanced/classes.html
-class PythonBoundaryConditionPythonSideInterfaceTrampoline
-    : public PythonBoundaryConditionPythonSideInterface
-{
-public:
-    using PythonBoundaryConditionPythonSideInterface::
-        PythonBoundaryConditionPythonSideInterface;
-
-    std::pair<bool, double> getDirichletBCValue(
-        double t, std::array<double, 3> x, std::size_t node_id,
-        std::vector<double> const& primary_variables) const override
-    {
-        using Ret = std::pair<bool, double>;
-        PYBIND11_OVERLOAD(Ret, PythonBoundaryConditionPythonSideInterface,
-                          getDirichletBCValue, t, x, node_id,
-                          primary_variables);
-    }
-
-    std::tuple<bool, double, std::vector<double>> getFlux(
-        double t, std::array<double, 3> x,
-        std::vector<double> const& primary_variables) const override
-    {
-        using Ret = std::tuple<bool, double, std::vector<double>>;
-        PYBIND11_OVERLOAD(Ret, PythonBoundaryConditionPythonSideInterface,
-                          getFlux, t, x, primary_variables);
-    }
 };
 }  // namespace ProcessLib
